update fix lỗi spin 2 v1

+++ app/src/main/java/com/example/khachhangarea_realfood/adapter/ThanhToanAdapter.java
@@ -75,7 +75,7 @@ public class ThanhToanAdapter extends RecyclerView.Adapter<ThanhToanAdapter.MyVi
             return;
         }
         holder.tvTenSanPham.setText(donHangInfo.getSanPham().getTenSanPham());
-        String gia = String.valueOf(Integer.valueOf(donHangInfo.getSanPham().getGia()));
+        String gia = String.valueOf(Double.valueOf(donHangInfo.getSanPham().getGia()));
         holder.tvGia.setText(gia);
         holder.tvSoLuong.setText("X "+ donHangInfo.getSoLuong());
         firebase_manager.storageRef.child("SanPham").child(donHangInfo.getSanPham().getIDCuaHang()).child(donHangInfo.getSanPham().getIDSanPham()).child(donHangInfo.getSanPham().getImages().get(0)).getDownloadUrl().addOnCompleteListener(new OnCompleteListener<Uri>() {
